app/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0

gunicorn==25.1.0

psycopg2-binary==2.9.11

python-dotenv==1.2.1

whitenoise==6.7.0
Pillow==10.4.0

django-filter==25.1
django-cors-headers==4.7.0

flake8==7.1.1
